ai/requirements.txt
numpy
onnxruntime
huggingface_hub
requests
torchvision
pillow
uvicorn
fastapi
pydantic
hf_xet
pydantic-settings
pyinstaller